[completion] Add doc comment for CodeCompletionSession
